+++ package.json
@@ -16,6 +16,7 @@
     "nekohack-ui": "2.9.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.0.2",
     "styled-components": "5.3.0",
     "swr": "^1.1.0"
   },
@@ -27,6 +28,7 @@
     "@types/jest": "26.0.24",
     "@types/react": "17.0.11",
     "@types/react-dom": "17.0.8",
+    "@types/react-router-dom": "5.3.2",
     "@types/react-test-renderer": "17.0.1",
     "@types/styled-components": "5.1.11",
     "@typescript-eslint/eslint-plugin": "4.28.0",
